+++ R/mus.R
@@ -24,14 +24,14 @@
 #' # Response
 #' y <- X %*% beta + rnorm(n, sd = 1)
 #' # Run the MU Selector
-#' mus1 <- mus(W, y)
+#' fit1 <- mus(W, y)
 #' # Draw an elbow plot to select delta
-#' plot(mus1)
+#' plot(fit1)
 #'
 #' # Now, according to the "elbow rule", choose the final delta where the curve has an "elbow".
 #' # In this case, the elbow is at about delta = 0.08, so we use this to compute the final estimate:
-#' mus2 <- mus(W, y, delta = 0.08)
-#' plot(mus2) # Plot the coefficients
+#' fit2 <- mus(W, y, delta = 0.08)
+#' plot(fit2) # Plot the coefficients
 #'
 #' @export
 mus <- function(W, y, lambda = NULL, delta = NULL) {